fix(test): MonitorDataProcessorTest ticks Int→Long to match MonitorPacket
Phase 6 changed MonitorPacket.ticks from Int to Long for tick overflow
prevention but missed updating the test helper's parameter type, causing
compileAndroidHostTest to fail on CI.

+++ shared/src/commonTest/kotlin/com/devil/phoenixproject/data/ble/MonitorDataProcessorTest.kt
@@ -30,7 +30,7 @@ class MonitorDataProcessorTest {
         posB: Float = 10.0f,
         loadA: Float = 5.0f,
         loadB: Float = 5.0f,
-        ticks: Int = 0,
+        ticks: Long = 0L,
         status: Int = 0,
         firmwareVelA: Int = 0,
         firmwareVelB: Int = 0,
